+++ Classes-MacOSX/UI/GHTestWindowController.m
@@ -45,6 +45,8 @@
 	self.window.contentView = viewController_.view;	
 	NSString *bundleVersion = [[NSBundle bundleForClass:[self class]] objectForInfoDictionaryKey:@"CFBundleVersion"];
 	self.window.title = [NSString stringWithFormat:@"GHUnit %@", bundleVersion];	
+  
+  if (getenv("GHUNIT_AUTORUN")) [self runTests:nil];
 }
 
 - (IBAction)runTests:(id)sender {