+++ Tests/GHSlowTest.m
@@ -14,10 +14,8 @@
 
 @implementation GHSlowTest
 
-#define kSlowTestTimeInterval 0.1
-
-- (void)testSlow {
-	[NSThread sleepForTimeInterval:5];
+- (void)test2Seconds {
+	[NSThread sleepForTimeInterval:2];
 }
 
 @end